pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "midcap-fortress"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.0",
  "pandas>=2.0",
  "numpy>=1.24",
  "yfinance>=0.2.0",
  "pyarrow>=14.0",
  "matplotlib>=3.8",
  "jinja2>=3.1",
  "rich>=13.0",
  "python-dotenv>=1.0",
  "alpaca-py>=0.27.0",
  "requests>=2.31",
]

[project.scripts]
mfp = "mfp.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6.0",
]
